+++ index.js
@@ -1,8 +1,15 @@
 module.exports = {
-    book: {
+    /*book: {
         assets: "./book",
         js: [
             "plugin.js"
         ]
+    },*/
+    hooks: {
+        'page': function(page) {
+            debugger;
+           console.log(arguments, this);
+            return page;
+        }
     }
 };
