+++ tests/replace.js
@@ -1,0 +1,8 @@
+var test = require('tape');
+var replace = require('./../src/replace');
+
+test('scanner', function (assert) {
+    replace();
+    assert.equal(true, true, 'replaced');
+    assert.end();
+});
